refactor : Notification Data 타입 추가 및 컴포넌트 분리
